[Add] some issue fixed

+++ database/migrations/2025_08_17_063733_create_packages_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->float('price')->default(0);
             $table->string('interest_type', 20)->default(INTEREST_TYPE_DAILY);
-            $table->integer('interest')->default(0);
+            $table->float('interest')->default(0);
             $table->date('duration')->default(now());
             $table->string('status', 20)->default(STATUS_ACTIVE);
             $table->timestamps();
